Init preview set view model
