+++ src/WinGetStudio/Views/ConfigurationFlow/PreviewFilePage.xaml.cs
@@ -65,7 +65,7 @@ public sealed partial class PreviewFilePage : Page, IView<PreviewFileViewModel>
             var selectedFile = await filePicker.ShowAsync(App.MainWindow);
             if (selectedFile != null)
             {
-                await ViewModel.OpenConfigurationFileAsync(selectedFile);
+                await ViewModel.Set.OpenConfigurationFileAsync(selectedFile);
             }
         }
         catch (Exception ex)
@@ -83,7 +83,7 @@ public sealed partial class PreviewFilePage : Page, IView<PreviewFileViewModel>
             var selectedFile = filePicker.Show(App.MainWindow);
             if (!string.IsNullOrEmpty(selectedFile))
             {
-                await ViewModel.SaveConfigurationAsAsync(selectedFile);
+                await ViewModel.Set.SaveConfigurationAsAsync(selectedFile);
             }
         }
         catch (Exception ex)
@@ -96,28 +96,28 @@ public sealed partial class PreviewFilePage : Page, IView<PreviewFileViewModel>
     {
         foreach (var id in e.AddedItems.OfType<UnitViewModel>())
         {
-            ViewModel.SelectedUnit?.Item2.Dependencies?.Add(id);
+            ViewModel.Set.SelectedUnit?.Item2.Dependencies?.Add(id);
         }
 
         foreach (var id in e.RemovedItems.OfType<UnitViewModel>())
         {
-            ViewModel.SelectedUnit?.Item2.Dependencies?.Remove(id);
+            ViewModel.Set.SelectedUnit?.Item2.Dependencies?.Remove(id);
         }
     }
 
     private void SelectedUnitDependencyLoaded(object sender, RoutedEventArgs e)
     {
-        if (sender is ListView listView && ViewModel.SelectedUnit != null)
+        if (sender is ListView listView && ViewModel.Set.SelectedUnit != null)
         {
             // Disable the option that matches the currently selected unit
             foreach (var item in listView.Items.OfType<UnitViewModel>())
             {
                 var container = listView.ContainerFromItem(item) as ListViewItem;
-                container?.IsEnabled = item != ViewModel.SelectedUnit.Item1;
+                container?.IsEnabled = item != ViewModel.Set.SelectedUnit.Item1;
             }
 
             // Get the set of IDs to select
-            var idsToSelect = ViewModel.SelectedUnit.Item2.Dependencies?.ToHashSet();
+            var idsToSelect = ViewModel.Set.SelectedUnit.Item2.Dependencies?.ToHashSet();
             if (idsToSelect == null || idsToSelect.Count == 0)
             {
                 listView.SelectedItems.Clear();
@@ -157,12 +157,12 @@ public sealed partial class PreviewFilePage : Page, IView<PreviewFileViewModel>
                 if (id == WinGetFileCodeLensHelper.EditResourceCommandId && value != null)
                 {
                     var index = value.Value.GetInt32();
-                    await ViewModel.OnEditUnitByIndexAsync(index);
+                    await ViewModel.Set.OnEditUnitByIndexAsync(index);
                 }
                 else if (arg?.Id == WinGetFileCodeLensHelper.ValidateUnitCommandId && value != null)
                 {
                     var index = value.Value.GetInt32();
-                    await ViewModel.OnValidateUnitByIndexAsync(index);
+                    await ViewModel.Set.OnValidateUnitByIndexAsync(index);
                 }
             }
             catch
@@ -174,17 +174,17 @@ public sealed partial class PreviewFilePage : Page, IView<PreviewFileViewModel>
 
     private void ViewModel_PropertyChanged(object? sender, PropertyChangedEventArgs e)
     {
-        if (e.PropertyName == nameof(ViewModel.Code))
+        if (e.PropertyName == nameof(ViewModel.Set.Code))
         {
-            var code = ViewModel.Code;
+            var code = ViewModel.Set.Code;
             if (!string.IsNullOrWhiteSpace(code) && WinGetFileCodeLensHelper.TryGenerateCodeLenses(_localizer, code, out var codeLenses))
             {
                 ConfigurationEditor.CodeLenses = codeLenses;
             }
         }
-        else if (e.PropertyName == nameof(ViewModel.IsCodeDirty))
+        else if (e.PropertyName == nameof(ViewModel.Set.IsCodeDirty))
         {
-            ConfigurationEditor.IsCodeLensEnabled = !ViewModel.IsCodeDirty;
+            ConfigurationEditor.IsCodeLensEnabled = !ViewModel.Set.IsCodeDirty;
         }
     }
 }
